# docs/chat/chat-opencode-integration.openapi.yaml
openapi: 3.1.0
info:
  title: OpenAWork Chat Opencode Integration API
  version: 0.1.0
paths:
...
  /sessions/{sessionId}/permissions/pending:
    get:
      summary: List pending permission requests for a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pending permission requests
  /sessions/{sessionId}/permissions/requests:
    post:
      summary: Create a permission request for a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Permission request created
  /sessions/{sessionId}/permissions/reply:
    post:
      summary: Reply to a permission request
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission reply accepted
...